base wait system
Version one of the wait system. The problem is that it do not wait long enough when the CSS appear to get the correct answer out of chat GPT so I need to user to tell me when to stop. But if the user tell it is over the script get the correct answer